added mock for electron-log and further separated renderer & main process code so that no modules are imported by both processes. #136. #141.
