+++ src/InvestJob.js
@@ -1,9 +1,7 @@
 import crypto from 'crypto';
 import localforage from 'localforage';
 
-import { getLogger } from './logger';
-
-const logger = getLogger(__filename.split('/').slice(-1)[0]);
+const logger = window.Workbench.getLogger(__filename.split('/').slice(-1)[0]);
 
 const HASH_ARRAY_KEY = 'workspaceHashes';
 const MAX_CACHED_JOBS = 30;
